pyproject.toml
[project]
name = "oiduna"
version = "0.1.0"
description = "Real-time SuperDirt/MIDI loop engine with HTTP API"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.112.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "python-osc>=1.8.3",
    "mido>=1.3.3",
    "python-rtmidi>=1.5.8",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.22",
]

[tool.uv]
package = false

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = [
    "packages/oiduna_api",
    "packages/oiduna_loop",
    "packages/oiduna_core",
    "packages/osc_protocol",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
